Using consistent naming
It is common in most Redlock implementations to call the update of the time to live of the lock `extension`.
However this is rather confusing, because implementations don't extend the time to live but rather overwrite it completely with a new value.
This results in the new TTL being shorter after using `extending` which I find rather unintuitive